uncertainty bug fixes and longs to int

+++ gui_gnome/PtCurMover.cpp
@@ -415,7 +415,7 @@ OSErr PtCurMover::PrepareForModelRun()
 	return noErr;
 }
 
-OSErr PtCurMover::PrepareForModelStep(const Seconds& model_time,const Seconds& time_step, bool uncertain, int numLESets, long* LESetsSizesList)
+OSErr PtCurMover::PrepareForModelStep(const Seconds& model_time,const Seconds& time_step, bool uncertain, int numLESets, int* LESetsSizesList)
 {
 	long timeDataInterval;
 	//Boolean intervalLoaded;
