+++ lib_gnome/CompoundMover_c.cpp
@@ -109,7 +109,7 @@ OSErr CompoundMover_c::PrepareForModelRun()
 	}
 	return noErr;
 }
-OSErr CompoundMover_c::PrepareForModelStep(const Seconds& model_time, const Seconds& time_step, bool uncertain, int numLESets, long* LESetsSizesList)
+OSErr CompoundMover_c::PrepareForModelStep(const Seconds& model_time, const Seconds& time_step, bool uncertain, int numLESets, int* LESetsSizesList)
 
 {
 	char errmsg[256];
